refactor: Card 관련 enum 패키지 변경
