+++ src/main/java/blackjack/domain/card/Card.java
@@ -2,8 +2,6 @@ package blackjack.domain.card;
 
 import static java.util.stream.Collectors.toSet;
 
-import blackjack.domain.Denomination;
-import blackjack.domain.Suit;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Objects;
