+++ src/test/java/blackjack/domain/DealerTest.java
@@ -1,9 +1,9 @@
 package blackjack.domain;
 
 import static blackjack.domain.CardsTestDataGenerator.*;
-import static blackjack.domain.Denomination.*;
+import static blackjack.domain.card.Denomination.*;
 import static blackjack.domain.GameResult.*;
-import static blackjack.domain.Suit.*;
+import static blackjack.domain.card.Suit.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
