+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,7 +1,7 @@
 package blackjack.domain.card;
 
-import static blackjack.domain.Denomination.*;
-import static blackjack.domain.Suit.*;
+import static blackjack.domain.card.Denomination.*;
+import static blackjack.domain.card.Suit.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Set;
