+++ src/test/java/blackjack/domain/card/CardsTest.java
@@ -1,8 +1,8 @@
 package blackjack.domain.card;
 
 import static blackjack.domain.CardsTestDataGenerator.generateCards;
-import static blackjack.domain.Denomination.*;
-import static blackjack.domain.Suit.*;
+import static blackjack.domain.card.Denomination.*;
+import static blackjack.domain.card.Suit.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
