+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,8 +1,8 @@
 package blackjack.domain.card;
 
 import static blackjack.domain.CardsTestDataGenerator.*;
-import static blackjack.domain.Denomination.*;
-import static blackjack.domain.Suit.*;
+import static blackjack.domain.card.Denomination.*;
+import static blackjack.domain.card.Suit.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
